editoast: move TrackSectionSncfExtension to editoast_schemas
