+++ editoast/editoast_schemas/src/infra/track_section_sncf_extension.rs
@@ -1,0 +1,16 @@
+use derivative::Derivative;
+use editoast_common::NonBlankString;
+use serde::Deserialize;
+use serde::Serialize;
+
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[serde(deny_unknown_fields)]
+#[derivative(Default)]
+pub struct TrackSectionSncfExtension {
+    pub line_code: i32,
+    #[derivative(Default(value = r#""line_test".into()"#))]
+    pub line_name: NonBlankString,
+    pub track_number: i32,
+    #[derivative(Default(value = r#""track_test".into()"#))]
+    pub track_name: NonBlankString,
+}
